refactor: split inline assets into css/app.css and js/app.js
Move the inline <style> and <script> blocks out of index.php into
web/css/app.css and web/js/app.js, leaving index.php as a slim 170-line
HTML template. Asset URLs include a filemtime() cache-buster so changes
appear without manual hard-reloads.

Add a print button in the bottom-right corner of the email meta-header
that calls iframe.contentWindow.print() to print only the email body.
